requirements.txt
cdp-sdk>=1.39.0
web3>=6.15.1
requests>=2.28.0
python-dotenv>=1.0.0
